and the restructuring continues
