+++ src/steps-trackers/qualifiers.ts
@@ -17,6 +17,8 @@
 import { Node as PMNode, Slice } from 'prosemirror-model'
 import { Selection, Transaction } from 'prosemirror-state'
 import { ReplaceAroundStep, ReplaceStep } from 'prosemirror-transform'
+import { ChangeSet } from '../ChangeSet'
+import { TrackedAttrs, CHANGE_OPERATION } from '../types/change'
 
 export const isDeleteStep = (step: ReplaceStep) =>
   step.from !== step.to && step.slice.content.size < step.to - step.from
@@ -114,4 +116,41 @@ export function isLiftStepForGap(
   to: number
 ) {
   return gap.start < gap.end && gap.insert === 0 && gap.end === to && !node.isText
+}
+
+/**
+ * Checks if this is a direct pending move deletion (not part of multiple moves)
+ *
+ * A direct pending move deletion occurs when:
+ * 1. The step is a deletion (from ≠ to and empty slice)
+ * 2. The step is not part of a larger move operation (not in movingSteps map)
+ * 3. The deleted node has pending move tracking attributes
+ *
+ * This is different from move operations that involve multiple steps (like cut-paste)
+ * where we need to track the relationship between deletion and insertion.
+ */
+export const isDirectPendingMoveDeletion = (
+  step: ReplaceStep,
+  doc: PMNode,
+  movingSteps: Map<ReplaceStep, string>
+): boolean => {
+  // Not a deletion
+  if (step.from === step.to || step.slice.content.size > 0) {
+    return false
+  }
+
+  // Part of a move operation
+  if (movingSteps.has(step)) {
+    return false
+  }
+
+  const node = doc.nodeAt(step.from)
+  if (!node) {
+    return false
+  }
+
+  return ChangeSet.isPendingChange(
+    node.attrs.dataTracked as TrackedAttrs[] | undefined,
+    CHANGE_OPERATION.move
+  )
 }
